requirements/requirements_3_1_2.txt
apache-airflow==3.1.2
apache-airflow-providers-slack
apache-airflow-providers-http
# apache-airflow-providers-standard  # for true HITL operators if desired
